🚚 Rename:폴더명도 PascalCase 기준으로 나누어 - 사용하는 것으로 통일 (#126)
