--- src/components/mypage/index.tsx
@@ -1,15 +1,0 @@
-import MyPageSideBar from '@/components/mypage/sidebar/MyPageSideBar'
-import { UserInfoDescription } from '@/components/mypage/myinfo/InfoDescription'
-import BookmarkedRecruitmentCard from '@/components/mypage/BookmarkedRecruitmentCard'
-import { PasswordChange } from '@/components/mypage/myinfo/PasswordChange'
-import { Withdrawal } from '@/components/mypage/myinfo/Withdrawal'
-import { InfoUpdate } from '@/components/mypage/myinfo/InfoUpdate'
-
-export {
-  MyPageSideBar,
-  UserInfoDescription,
-  BookmarkedRecruitmentCard,
-  PasswordChange,
-  Withdrawal,
-  InfoUpdate,
-}
