+++ src/pages/my-page/BookmarkedRecruitment.tsx
@@ -1,6 +1,6 @@
 import { ListItemSkeleton } from '@/components'
 import { Input, InputIcon } from '@/components/common/input'
-import { BookmarkedRecruitmentCard } from '@/components/mypage'
+import { BookmarkedRecruitmentCard } from '@/components/my-page'
 import EmptyDataState from '@/components/common/State/EmptyDataState'
 import { useBookmarkedRecruitment } from '@/hooks/api'
 import { useDebounce } from '@/hooks/useDebounce'
